requirements.txt
cycler==0.10.0
dask>=2.17.2
decorator>=4.3.0
docopt==0.6.2
ijson==2.3
matplotlib>=2.0.0
networkx>=2.0
numpy>=1.11.1
Pillow>=4.3.0
pyparsing==2.1.5
python-dateutil>=2.6.1
pytz>=2017.2
scikit-image>=0.12.3
scikit-learn>=0.17.1
scipy>=0.17.1
six==1.11.0
toolz>=0.8.0
tqdm>=4.46.1
webcolors>=1.11.1
streamlit>=0.60.0
fastapi==0.58.0
uvicorn==0.11.5
python-multipart==0.0.5
gunicorn==20.0.4
starlette-exporter==0.4.1
aiofiles==0.5.0
